refactor: replace student editor sheet with dialog
