+++ app/src/main/java/com/tutorly/ui/screens/StudentDetailsScreen.kt
@@ -46,7 +46,6 @@ import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.IconButtonDefaults
-import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.SnackbarHost
@@ -76,12 +75,10 @@ import androidx.hilt.navigation.compose.hiltViewModel
 import com.tutorly.R
 import com.tutorly.domain.model.StudentProfile
 import com.tutorly.domain.model.StudentProfileLesson
-import com.tutorly.models.SubjectPreset
 import com.tutorly.models.PaymentStatus
 import com.tutorly.ui.components.TopBarContainer
 import com.tutorly.ui.components.PaymentBadge
 import com.tutorly.ui.components.PaymentBadgeStatus
-import com.tutorly.ui.components.TutorlyDialog
 import com.tutorly.ui.theme.extendedColors
 import com.tutorly.ui.lessoncard.LessonCardSheet
 import com.tutorly.ui.lessoncard.LessonCardViewModel
@@ -558,105 +555,6 @@ private fun StudentProfileTopBar(
                 }
             }
         }
-    }
-}
-
-@Composable
-private fun StudentEditorDialogContent(
-    state: StudentEditorFormState,
-    onNameChange: (String) -> Unit,
-    onPhoneChange: (String) -> Unit,
-    onMessengerChange: (String) -> Unit,
-    onRateChange: (String) -> Unit,
-    subjectPresets: List<SubjectPreset>,
-    onSubjectChange: (String) -> Unit,
-    onGradeChange: (String) -> Unit,
-    onNoteChange: (String) -> Unit,
-    onSave: () -> Unit,
-    onDismiss: () -> Unit,
-    editTarget: StudentEditTarget?,
-    initialFocus: StudentEditTarget?,
-    snackbarHostState: SnackbarHostState,
-) {
-    val titleRes = when {
-        state.studentId == null -> R.string.add_student
-        editTarget == StudentEditTarget.RATE -> R.string.student_editor_title_rate
-        editTarget == StudentEditTarget.PHONE -> R.string.student_editor_title_phone
-        editTarget == StudentEditTarget.MESSENGER -> R.string.student_editor_title_messenger
-        editTarget == StudentEditTarget.NOTES -> R.string.student_editor_title_note
-        else -> R.string.student_editor_title
-    }
-    val title = stringResource(id = titleRes)
-
-    TutorlyDialog(
-        onDismissRequest = {
-            if (!state.isSaving) {
-                onDismiss()
-            }
-        },
-        modifier = Modifier.imePadding()
-    ) {
-        Text(
-            text = title,
-            style = MaterialTheme.typography.titleLarge
-        )
-
-        if (state.isSaving) {
-            LinearProgressIndicator(
-                modifier = Modifier.fillMaxWidth(),
-                color = MaterialTheme.colorScheme.secondary
-            )
-        }
-
-        StudentEditorForm(
-            state = state,
-            onNameChange = onNameChange,
-            onPhoneChange = onPhoneChange,
-            onMessengerChange = onMessengerChange,
-            onRateChange = onRateChange,
-            subjectPresets = subjectPresets,
-            onSubjectChange = onSubjectChange,
-            onGradeChange = onGradeChange,
-            onNoteChange = onNoteChange,
-            modifier = Modifier.fillMaxWidth(),
-            editTarget = editTarget,
-            initialFocus = initialFocus,
-            enableScrolling = true,
-            enabled = !state.isSaving,
-            onSubmit = onSave
-        )
-
-        Row(
-            modifier = Modifier.fillMaxWidth(),
-            horizontalArrangement = Arrangement.End,
-            verticalAlignment = Alignment.CenterVertically
-        ) {
-            val accent = MaterialTheme.extendedColors.accent
-            val actionColors = ButtonDefaults.textButtonColors(
-                contentColor = accent,
-                disabledContentColor = accent.copy(alpha = 0.5f)
-            )
-            TextButton(
-                onClick = onDismiss,
-                enabled = !state.isSaving,
-                colors = actionColors
-            ) {
-                Text(text = stringResource(id = R.string.student_editor_cancel))
-            }
-            Spacer(modifier = Modifier.width(8.dp))
-            TextButton(
-                onClick = onSave,
-                enabled = !state.isSaving && state.name.isNotBlank(),
-                colors = actionColors
-            ) {
-                Text(text = stringResource(id = R.string.student_editor_save))
-            }
-        }
-
-        SnackbarHost(
-            hostState = snackbarHostState,
-            modifier = Modifier.fillMaxWidth()
-        )
     }
 }
